cli decoupling wip
